Change toolbarView for taking handlers by xml-attributes
