+++ presentation/src/main/java/by/grodno/vasili/presentation/feature/notedetails/DetailsActivity.java
@@ -43,47 +43,13 @@ public class DetailsActivity extends BaseActivity<ActivityDetailsBinding> {
         super.onCreate(savedInstanceState);
         model = ViewModelProviders.of(this, provider).get(DetailsViewModel.class);
         binding.setHandler(new Handler());
-        initToolbar();
         loadNoteAsync();
         model.noteLiveData.observe(this, noteItem -> binding.setNote(noteItem));
     }
 
-    private void initToolbar() {
-        binding.toolbarView.setOkListener(view -> saveIfChanged());
-    }
-
-    /**
-     * Checking entered data for correct and presence of changes and save
-     */
-    private void saveIfChanged() {
-        NoteItem item = binding.getNote();
-        String changedTitle = trim(binding.titleText.getText().toString());
-        String changedDescription = trim(binding.descriptionText.getText().toString());
-        if (!validateTitle(changedTitle) || !validateDescription(changedDescription)) {
-            showToast(String.format(Locale.US,
-                    "Entered data can`t be empty. Maximum length for title: %d, for description: %d",
-                    MAX_TITLE_LENGTH,
-                    MAX_DESCRIPTION_LENGTH));
-            return;
-        }
-        NoteItem changedItem = new NoteItem(item.id, changedTitle, changedDescription, item.created);
-        if (hasChanges(item, changedItem)) {
-            Consumer<String> onSuccess = id -> showToast(String.format("Note %s successfully updated", id));
-            Consumer<Throwable> onError = error -> showToast("Error updating note: " + error.getLocalizedMessage());
-            model.save(changedItem, onSuccess, onError);
-            notifyAboutChanges();
-        }
-        finish();
-    }
-
-    private void notifyAboutChanges() {
-        Intent intent = new Intent();
-        intent.putExtra(NOTE_CHANGED, true);
-        setResult(RESULT_OK, intent);
-    }
-
-    private boolean hasChanges(NoteItem original, NoteItem changed) {
-        return !original.title.equals(changed.title) || !original.description.equals(changed.description);
+    @Override
+    protected int getContentView() {
+        return R.layout.activity_details;
     }
 
     private void loadNoteAsync() {
@@ -94,11 +60,6 @@ public class DetailsActivity extends BaseActivity<ActivityDetailsBinding> {
         } else {
             model.getNoteAsync(id);
         }
-    }
-
-    @Override
-    protected int getContentView() {
-        return R.layout.activity_details;
     }
 
     /**
@@ -113,6 +74,40 @@ public class DetailsActivity extends BaseActivity<ActivityDetailsBinding> {
             EditText descriptionView = binding.descriptionText;
             enableEditing(titleView, descriptionView);
             binding.editSwitch.setVisibility(View.GONE);
+        }
+
+        /**
+         * Checking entered data for correct and presence of changes and save
+         */
+        public void saveIfChanged() {
+            NoteItem item = binding.getNote();
+            String changedTitle = trim(binding.titleText.getText().toString());
+            String changedDescription = trim(binding.descriptionText.getText().toString());
+            if (!validateTitle(changedTitle) || !validateDescription(changedDescription)) {
+                showToast(String.format(Locale.US,
+                        "Entered data can`t be empty. Maximum length for title: %d, for description: %d",
+                        MAX_TITLE_LENGTH,
+                        MAX_DESCRIPTION_LENGTH));
+                return;
+            }
+            NoteItem changedItem = new NoteItem(item.id, changedTitle, changedDescription, item.created);
+            if (hasChanges(item, changedItem)) {
+                Consumer<String> onSuccess = id -> showToast(String.format("Note %s successfully updated", id));
+                Consumer<Throwable> onError = error -> showToast("Error updating note: " + error.getLocalizedMessage());
+                model.save(changedItem, onSuccess, onError);
+                notifyAboutChanges();
+            }
+            finish();
+        }
+
+        private void notifyAboutChanges() {
+            Intent intent = new Intent();
+            intent.putExtra(NOTE_CHANGED, true);
+            setResult(RESULT_OK, intent);
+        }
+
+        private boolean hasChanges(NoteItem original, NoteItem changed) {
+            return !original.title.equals(changed.title) || !original.description.equals(changed.description);
         }
 
         private void enableEditing(EditText... views) {
